+++ CHANGELOG.md
@@ -25,7 +25,7 @@ fix: initiate `id` for tables as UNSIGNED (not left as default, which is SIGNED)
 
 ### Added
 
-- MailOut method to send emails
+- MailOut method to send emails (Sending of emails to users MUST be activated, so that `$this->configuration->flag->status(SeablastConstant::USER_MAIL_ENABLED)` is true.)
 
 ### Fixed